HUB-595: replace live with variant
This step replace current live with the variant. Renaming the variant
to the live versions.
