+++ config/main_routes.rb
@@ -45,24 +45,33 @@ constraints IsLoa2 do
   #  post "start", to: "start#request_post", as: :start
   #  get "begin_registration", to: "start#register", as: :begin_registration
   #end
+  #constraints sign_in_hint.use(alternative: "variant") do
+  #  get "prove_identity", to: "prove_identity_variant#index", as: :prove_identity
+  #  get "prove_identity_retry", to: "prove_identity_variant#retry_eidas_journey", as: :prove_identity_retry
+  #  get "prove_identity_ignore_hint", to: "prove_identity_variant#ignore_hint", as: :prove_identity_ignore_hint
+  #  get "start", to: "start_variant#index", as: :start
+  #  post "start", to: "start_variant#request_post", as: :start
+  #  get "start_ignore_hint", to: "start_variant#ignore_hint", as: :start_ignore_hint
+  #  get "begin_registration", to: "start_variant#register", as: :begin_registration
+  #end
   constraints sign_in_hint.use(alternative: "control") do
-    get "prove_identity", to: "prove_identity_variant#index", as: :prove_identity
-    get "prove_identity_retry", to: "prove_identity_variant#retry_eidas_journey", as: :prove_identity_retry
-    get "prove_identity_ignore_hint", to: "prove_identity_variant#ignore_hint", as: :prove_identity_ignore_hint
-    get "start", to: "start_variant#index", as: :start
-    post "start", to: "start_variant#request_post", as: :start
-    get "start_ignore_hint", to: "start_variant#ignore_hint", as: :start_ignore_hint
-    get "begin_registration", to: "start_variant#register", as: :begin_registration
+    get "prove_identity", to: "prove_identity#index", as: :prove_identity
+    get "prove_identity_retry", to: "prove_identity#retry_eidas_journey", as: :prove_identity_retry
+    get "prove_identity_ignore_hint", to: "prove_identity#ignore_hint", as: :prove_identity_ignore_hint
+    get "start", to: "start#index", as: :start
+    post "start", to: "start#request_post", as: :start
+    get "start_ignore_hint", to: "start#ignore_hint", as: :start_ignore_hint
+    get "begin_registration", to: "start#register", as: :begin_registration
   end
 
   constraints sign_in_hint.use(alternative: "variant") do
-    get "prove_identity", to: "prove_identity_variant#index", as: :prove_identity
-    get "prove_identity_retry", to: "prove_identity_variant#retry_eidas_journey", as: :prove_identity_retry
-    get "prove_identity_ignore_hint", to: "prove_identity_variant#ignore_hint", as: :prove_identity_ignore_hint
-    get "start", to: "start_variant#index", as: :start
-    post "start", to: "start_variant#request_post", as: :start
-    get "start_ignore_hint", to: "start_variant#ignore_hint", as: :start_ignore_hint
-    get "begin_registration", to: "start_variant#register", as: :begin_registration
+    get "prove_identity", to: "prove_identity#index", as: :prove_identity
+    get "prove_identity_retry", to: "prove_identity#retry_eidas_journey", as: :prove_identity_retry
+    get "prove_identity_ignore_hint", to: "prove_identity#ignore_hint", as: :prove_identity_ignore_hint
+    get "start", to: "start#index", as: :start
+    post "start", to: "start#request_post", as: :start
+    get "start_ignore_hint", to: "start#ignore_hint", as: :start_ignore_hint
+    get "begin_registration", to: "start#register", as: :begin_registration
   end
 
   # get "prove_identity", to: "prove_identity#index", as: :prove_identity
