+++ spec/controllers/start_controller_spec.rb
@@ -79,4 +79,76 @@ describe StartController do
     expect(subject).to redirect_to("/about")
     expect(stub_piwik_request).to have_been_made.once
   end
+
+  context "when sign-in hint is present" do
+    it "renders the hint when IDP valid" do
+      cookies.encrypted[CookieNames::VERIFY_FRONT_JOURNEY_HINT] = {
+        "SUCCESS" => "http://idcorp.com",
+      }.to_json
+      stub_api_idp_list_for_sign_in
+
+      get :index, params: { locale: "en" }
+      expect(subject).to render_template("shared/sign_in_hint", "layouts/main_layout")
+    end
+
+    it "renders the normal start page if IDP is invalid" do
+      cookies.encrypted[CookieNames::VERIFY_FRONT_JOURNEY_HINT] = {
+        "SUCCESS" => "invalid",
+      }.to_json
+      stub_api_idp_list_for_sign_in
+
+      get :index, params: { locale: "en" }
+      expect(subject).to render_template(:start, "layouts/slides")
+    end
+
+    it "renders the normal start page if success is missing" do
+      cookies.encrypted[CookieNames::VERIFY_FRONT_JOURNEY_HINT] = {
+        "ATTEMPT" => "http://idcorp.com",
+      }.to_json
+      stub_api_idp_list_for_sign_in
+
+      get :index, params: { locale: "en" }
+      expect(subject).to render_template(:start, "layouts/slides")
+    end
+
+    it "allows to disregard the hint and deletes the SUCCESS" do
+      cookies.encrypted[CookieNames::VERIFY_FRONT_JOURNEY_HINT] = {
+        "ATTEMPT" => "http://idcorp.com",
+        "SUCCESS" => "http://idcorp.com",
+      }.to_json
+      stub_api_idp_list_for_sign_in
+
+      expect(FEDERATION_REPORTER).to receive(:report_sign_in_journey_ignored).with(
+        a_kind_of(Display::RpDisplayData),
+        a_kind_of(ActionDispatch::Request),
+        "IDCorp",
+        "test-rp",
+      )
+
+      get :ignore_hint, params: { locale: "en" }
+
+      cookie_hint = MultiJson.load(cookies.encrypted[CookieNames::VERIFY_FRONT_JOURNEY_HINT])
+
+      expect(subject).to redirect_to start_path
+      expect(cookie_hint["ATTEMPT"]).to eq "http://idcorp.com"
+      expect(cookie_hint["SUCCESS"]).to be_nil
+    end
+
+    it "allows to disregard the hint and does not fail if success does not exist" do
+      cookies.encrypted[CookieNames::VERIFY_FRONT_JOURNEY_HINT] = {
+        "ATTEMPT" => "http://idcorp.com",
+      }.to_json
+      stub_api_idp_list_for_sign_in
+
+      expect(FEDERATION_REPORTER).not_to receive(:report_sign_in_journey_ignored)
+
+      get :ignore_hint, params: { locale: "en" }
+
+      cookie_hint = MultiJson.load(cookies.encrypted[CookieNames::VERIFY_FRONT_JOURNEY_HINT])
+
+      expect(subject).to redirect_to start_path
+      expect(cookie_hint["ATTEMPT"]).not_to be_nil
+      expect(cookie_hint["SUCCESS"]).to be_nil
+    end
+  end
 end
