+++ spec/features/user_visits_start_page_spec.rb
@@ -3,25 +3,56 @@ require "api_test_helper"
 require "cookie_names"
 
 RSpec.describe "When the user visits the start page" do
-  it "will display the start page in English" do
-    set_session_and_session_cookies!
-    visit "/start"
-    expect(page).to have_content t("hub.start.heading")
-    expect(page).to have_css "html[lang=en]"
-    expect_feedback_source_to_be(page, "START_PAGE", "/start")
-  end
+  context "when session is valid" do
+    before(:each) do
+      set_session_and_ab_session_cookies!("sign_in_hint" => "sign_in_hint_variant")
+    end
 
-  it "will display the start page in Welsh" do
-    set_session_and_session_cookies!
-    visit "/dechrau"
-    expect(page).to have_content t("hub.start.heading", locale: :cy)
-    expect(page).to have_css "html[lang=cy]"
-  end
+    it "will display the start page in English on variant" do
+      set_session_and_session_cookies!
+      visit "/start"
+      expect(page).to have_content t("hub.start.heading")
+      expect(page).to have_css "html[lang=en]"
+      expect_feedback_source_to_be(page, "START_PAGE", "/start")
+    end
 
-  it "will not automatically disable the continue button on submit" do
-    set_session_and_session_cookies!
-    visit "/start"
-    expect(page).to_not have_css "#next-button[data-disable-with]"
+    it "will display the hint page if an success hint present" do
+      set_session_and_session_cookies!
+      set_journey_hint_cookie("http://idcorp.com", "SUCCESS")
+      stub_api_idp_list_for_sign_in
+      visit "/start"
+      expect(page).to have_content t("hub.sign_in_hint.heading")
+      expect(page).to have_css "html[lang=en]"
+    end
+
+    it "will reset the hint and display start page when user ignores the hint" do
+      set_session_and_session_cookies!
+      set_journey_hint_cookie("http://idcorp.com", "SUCCESS")
+      stub_api_idp_list_for_sign_in
+      visit "/start"
+      expect(page).to have_content t("hub.sign_in_hint.heading")
+      expect(page).to have_css "html[lang=en]"
+      expect(page).to have_current_path "/start"
+
+      click_link t("hub.sign_in_hint.other_way_button")
+
+      expect(page).to have_content t("hub.start.heading")
+      expect(page).to have_css "html[lang=en]"
+      expect(page).to have_current_path "/start"
+    end
+
+    it "will display the start page in Welsh" do
+      set_session_and_session_cookies!
+      visit "/dechrau"
+      expect(page).to have_content t("hub.start.heading", locale: :cy)
+      expect(page).to have_css "html[lang=cy]"
+    end
+
+    it "will not automatically disable the continue button on submit" do
+      set_session_and_session_cookies!
+      visit "/start"
+      expect(page).to_not have_css "#next-button[data-disable-with]"
+    end
   end
 
   context "when there is an error" do
@@ -105,7 +136,6 @@ RSpec.describe "When the user visits the start page" do
       expect(page.body).to include t("errors.session_timeout.return_to_service_html")
       expect(page).to have_http_status :forbidden
       expect(page).to have_link "feedback", href: "/feedback-landing?feedback-source=EXPIRED_ERROR_PAGE"
-      expect(page).to have_link t("errors.session_timeout.start_again"), href: "http://www.test-rp.gov.uk/"
     end
   end
